pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kraft"
version = "0.1.0"
description = "Game Creation OS — crie jogos com IA, teste no navegador, venda na loja"
requires-python = ">=3.12"
dependencies = [
    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.4",
    "aiofiles>=23.2.1",
    "websockets>=12.0",

    # Banco de dados
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.29.0",
    "alembic>=1.13.3",

    # Validação e settings
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.2",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Redis e filas
    "redis>=5.1.0",
    "dramatiq[redis]>=1.17.0",

    # Storage (S3/R2)
    "boto3>=1.35.0",

    # HTTP client (para IA)
    "httpx>=0.27.0",
    "anthropic>=0.36.0",

    # Utilitários
    "python-slugify>=8.0.4",
    "pillow>=10.4.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
